fix: lint warnings with context
