+++ gbajs3/src/context/emulator/contexts/drag-context.tsx
@@ -1,0 +1,12 @@
+import { createContext } from 'react';
+
+import type { Dispatch, SetStateAction } from 'react';
+
+type DragContextProps = {
+  areItemsDraggable: boolean;
+  setAreItemsDraggable: Dispatch<SetStateAction<boolean>>;
+};
+
+export const DragContext = createContext<DragContextProps | null>(null);
+
+DragContext.displayName = 'DragContext';
