+++ gbajs3/src/context/emulator/providers/drag-provider.tsx
@@ -1,0 +1,17 @@
+import { useState, type ReactNode } from 'react';
+
+import { DragContext } from '../contexts/drag-context.tsx';
+
+type DragProviderProps = {
+  children: ReactNode;
+};
+
+export const DragProvider = ({ children }: DragProviderProps) => {
+  const [areItemsDraggable, setAreItemsDraggable] = useState(false);
+
+  return (
+    <DragContext.Provider value={{ areItemsDraggable, setAreItemsDraggable }}>
+      {children}
+    </DragContext.Provider>
+  );
+};
