+++ gbajs3/src/hooks/context.spec.tsx
@@ -10,15 +10,15 @@ import {
   useResizeContext,
   useRunningContext
 } from './context.tsx';
-import { AuthContext } from '../context/auth/auth.tsx';
-import { DragContext } from '../context/emulator/drag.tsx';
-import { EmulatorContext } from '../context/emulator/emulator.tsx';
-import { ResizeContext } from '../context/emulator/resize.tsx';
-import { RunningContext } from '../context/emulator/running.tsx';
-import { LayoutContext } from '../context/layout/layout.tsx';
-import { ModalContext } from '../context/modal/modal.tsx';
+import { AuthContext } from '../context/auth/auth-context.tsx';
+import { DragContext } from '../context/emulator/contexts/drag-context.tsx';
+import { EmulatorContext } from '../context/emulator/contexts/emulator-context.tsx';
+import { ResizeContext } from '../context/emulator/contexts/resize-context.tsx';
+import { RunningContext } from '../context/emulator/contexts/running-context.tsx';
+import { LayoutContext } from '../context/layout/layout-context.tsx';
+import { ModalContext } from '../context/modal/modal-context.tsx';
 
-import type * as authContextExports from '../context/auth/auth.tsx';
+import type * as authContextExports from '../context/auth/auth-context.tsx';
 
 describe('useContext hooks', () => {
   beforeEach(() => {
@@ -53,10 +53,10 @@ describe('useContext hooks', () => {
     // silence console errors as they are expected
     vi.spyOn(console, 'error').mockImplementation(() => {});
 
-    vi.doMock('../context/auth/auth.tsx', async () => {
+    vi.doMock('../context/auth/auth-context.tsx', async () => {
       const { AuthContext: original, ...rest } = await vi.importActual<
         typeof authContextExports
-      >('../context/auth/auth.tsx');
+      >('../context/auth/auth-context.tsx');
 
       return {
         ...rest,
