+++ gbajs3/test/providers.tsx
@@ -1,9 +1,9 @@
 import { ThemeProvider } from 'styled-components';
 
-import { AuthProvider } from '../src/context/auth/auth.tsx';
+import { AuthProvider } from '../src/context/auth/auth-provider.tsx';
 import { EmulatorContextProvider } from '../src/context/emulator/emulator-context-provider.tsx';
-import { LayoutProvider } from '../src/context/layout/layout.tsx';
-import { ModalProvider } from '../src/context/modal/modal.tsx';
+import { LayoutProvider } from '../src/context/layout/layout-provider.tsx';
+import { ModalProvider } from '../src/context/modal/modal-provider.tsx';
 import { GbaDarkTheme } from '../src/context/theme/theme.tsx';
 
 import type { ReactNode } from 'react';
